[chore]: api 파일 경로들 수정 (#22)
