+++ src/api/hooks/home/useCalendars.ts
@@ -1,19 +1,17 @@
 import { useState, useEffect } from "react";
-import useAuthStore from "@/stores/authStore";
-import { getCalendars, type CalendarItem } from "@/types/home";
+import type { CalendarItem, CalendarResponse } from "@/types/home";
+import { http } from "@/api/fetch";
+
+export const getCalendars = async (): Promise<CalendarResponse> => {
+  const { response } = await http.get<CalendarResponse>(`/api/v1/calendars`);
+  return response;
+};
 
 export const useCalendars = () => {
-  const { isLoggedIn } = useAuthStore();
-
   const [calendars, setCalendars] = useState<CalendarItem[]>([]);
 
   useEffect(() => {
     const fetchCalendars = async () => {
-      if (!isLoggedIn) {
-        setCalendars([]); // 로그아웃 시 데이터 초기화
-        return;
-      }
-
       try {
         const data = await getCalendars();
         setCalendars(data);
@@ -23,7 +21,7 @@ export const useCalendars = () => {
     };
 
     fetchCalendars();
-  }, [isLoggedIn]); // 로그인 상태가 변경될 때마다 데이터를 다시 가져옵니다.
+  }, []);
 
   return { calendars };
 };
