+++ src/api/hooks/notice/useNotices.ts
@@ -1,11 +1,34 @@
+import { http } from "@/api/fetch";
 import {
-  getNotices,
-  addBookmark,
-  deleteBookmark,
   type GetNoticesParams,
   type NoticeData,
+  type NoticeResponse,
 } from "@/types/notice";
 import { useState, useEffect, useCallback } from "react";
+
+export const getNotices = async (
+  params: GetNoticesParams
+): Promise<NoticeResponse> => {
+  const cleanParams = Object.fromEntries(
+    Object.entries(params).filter(
+      ([, value]) => value !== undefined && value !== ""
+    )
+  ) as Record<string, string>;
+
+  const { response } = await http.get<NoticeResponse>(
+    `/api/v1/notices`,
+    cleanParams
+  );
+  return response;
+};
+
+export const addBookmark = async (noticeId: number): Promise<void> => {
+  await http.post(`/api/v1/notices/${noticeId}/bookmark`);
+};
+
+export const deleteBookmark = async (noticeId: number): Promise<void> => {
+  await http.delete(`/api/v1/notices/${noticeId}/bookmark`);
+};
 
 export const useNotices = () => {
   const [params, setParams] = useState<GetNoticesParams>({
