+++ src/types/notice.ts
@@ -1,5 +1,3 @@
-import { http } from "@/api/fetch";
-
 export interface NoticeData {
   id: number;
   categoryId: number;
@@ -27,27 +25,3 @@ export interface NoticeResponse {
   totalPages: number;
   last: boolean;
 }
-
-export const getNotices = async (
-  params: GetNoticesParams
-): Promise<NoticeResponse> => {
-  const cleanParams = Object.fromEntries(
-    Object.entries(params).filter(
-      ([, value]) => value !== undefined && value !== ""
-    )
-  ) as Record<string, string>;
-
-  const { response } = await http.get<NoticeResponse>(
-    `/api/v1/notices`,
-    cleanParams
-  );
-  return response;
-};
-
-export const addBookmark = async (noticeId: number): Promise<void> => {
-  await http.post(`/api/v1/notices/${noticeId}/bookmark`);
-};
-
-export const deleteBookmark = async (noticeId: number): Promise<void> => {
-  await http.delete(`/api/v1/notices/${noticeId}/bookmark`);
-};
